xwayland: fix EnterNotify position
enter handler wasn't updating sprite coordinates based on the Wayland event
just sent, failing when forwarding the correct EnterNotify to the X client.

There was a doubt regarding whether ProcessInputEvents should be called just
after SetCursorPosition or not in order to dequeue the event. When the cursor
is in motion it obviously will trigger PIE, so to verify it without moving I
used the switcher binding (mod + tab), so the cursor would enter another
(frame) window that was unmapped, change to the right cursor and PIE would
have to be triggered right after. In fact that was the case and I simply
assumed a direct call to PIE wasn't really needed there.

Signed-off-by: Tiago Vignatti <tiago.vignatti@intel.com>